RS-33 RS-34 improve scanner management, CLI options, and cleanup logic
- Adds new CLI options to control macOS filesystem case sensitivity.
- Refactors and enhances the ScannerManager with improved resource management, mutex handling, and cleanup logic.
- Introduces more robust path normalization and deduplication for repositories, including case sensitivity handling.
- Improves error handling and retry policies for scanner and queue operations.
- Updates application startup and shutdown coordination for better reliability and graceful termination.
- Removes inline test modules from several files and deletes the unused plugin/context.rs.
- Reorganizes CLI module structure and adds new submodules for config, parsing, and validation.
