+++ src/app/cli/tests/mod.rs
@@ -1,0 +1,9 @@
+//! Tests for the CLI module
+//!
+//! This module contains all tests for CLI argument parsing, validation,
+//! and related functionality, extracted from individual modules for better organization.
+
+pub mod args_tests;
+pub mod command_segmenter_tests;
+pub mod date_parser_tests;
+pub mod display_tests;
