+++ src/plugin/mod.rs
@@ -6,14 +6,12 @@
 // Internal modules - all access should go through api module
 pub(crate) mod args;
 pub(crate) mod builtin;
-pub(crate) mod context;
 pub(crate) mod discovery;
 pub(crate) mod error;
 pub(crate) mod error_handling;
 pub(crate) mod events;
 pub(crate) mod manager;
 pub(crate) mod registry;
-pub(crate) mod settings;
 pub(crate) mod traits;
 pub(crate) mod types;
 pub(crate) mod unified_discovery;
